Put the key in the index; widen visit counts to u32
Three changes to the node stack, all measured.

The key index stores the key beside the node slot, so probing never touches a
node. It was dereferencing one on every occupied slot just to compare a key, and
nodes are ~2 KB in an array approaching a gigabyte, so that was a likely cache
miss per probe step on the hottest path in the search. At 200k nodes and 411 MB:

    lookup, hit    34.4 -> 14.4 ns    2.4x
    lookup, miss   21.0 -> 11.8 ns    1.8x

Index slots go 4 -> 16 bytes, four to a cache line, 16 MB against ~1 GB of nodes.

Deletion still tombstones. Shifting later entries back into the hole would remove
tombstones and rehashing entirely, and it fits a fixed-capacity table whose values
are indices — but it moves entries, which makes the table hostile to concurrent
readers, and 800k simulations per step will want threads. Tombstones are cheap now
anyway: walking past one is an in-cache read rather than a node dereference.

The rehash trigger was wrong. It fired on tombstones alone, so live entries plus
tombstones could reach 0.75 of the index and every lookup would walk a long run.
It now fires on total occupancy, which is what probe length actually depends on.
probe_lengths() reports mean and worst on demand, walked rather than counted per
lookup, and a churn test holds the table near capacity for 300 rounds: 21
rebuilds, mean probe 1.62, worst 18.

Visit counts are u32. The u16 ceiling was reachable in principle because a shared
node collects visits from every game that reaches it; at inference scale it may
stop being theoretical. The saturation counter stays as a tripwire.

+++ rust/src/mcts/variant.rs
@@ -90,9 +90,10 @@ fn count(legal: [u64; LEGAL_WORDS]) -> u32 {
 #[derive(Clone, Debug)]
 pub struct PuctStats {
     pub prior: [[f32; SQUARES]; 2],
-    pub visit: [[u16; SQUARES]; 2],
+    pub visit: [[u32; SQUARES]; 2],
     pub q: [[f32; SQUARES]; 2],
-    /// Backups dropped because an edge hit the `u16` visit ceiling.
+    /// Backups dropped because an edge hit the `u32` visit ceiling. Unreachable in practice;
+    /// it is here so that if it ever happens it is visible rather than silent.
     pub saturated: u32,
 }
 
@@ -140,13 +141,13 @@ impl Variant for Puct {
         _rng: &mut Rng,
     ) -> Choice {
         let (q, visit, prior) = (&stats.q[player], &stats.visit[player], &stats.prior[player]);
-        let total: u32 = squares(legal).map(|sq| u32::from(visit[sq])).sum();
+        let total: u64 = squares(legal).map(|sq| u64::from(visit[sq])).sum();
         let explore = cfg.c_puct * (total as f32).sqrt();
 
         let mut best = f32::NEG_INFINITY;
         let mut action = usize::MAX;
         for sq in squares(legal) {
-            let score = q[sq] + explore * prior[sq] / (1.0 + f32::from(visit[sq]));
+            let score = q[sq] + explore * prior[sq] / (1.0 + visit[sq] as f32);
             if score > best {
                 best = score;
                 action = sq;
@@ -166,12 +167,12 @@ impl Variant for Puct {
         for player in 0..2 {
             let sq = choice[player].action as usize;
             let n = stats.visit[player][sq];
-            if n == u16::MAX {
+            if n == u32::MAX {
                 stats.saturated += 1;
                 continue;
             }
             let q = &mut stats.q[player][sq];
-            *q = (*q * f32::from(n) + values[player]) / (f32::from(n) + 1.0);
+            *q = (*q * n as f32 + values[player]) / (n as f32 + 1.0);
             stats.visit[player][sq] = n + 1;
         }
     }
